Add functions to support SDES
These functions create and parse the SDES crypto string and create the
corresponding SRTP profiles. An application uses the SDES strings in
SDP to exchange key data and the name of the crypto suite.

The application uses the SRTP profiles to set up the SRTP functions
to protect and unprotect RTP traffic.

Modified-by: Travis Cross <tc@traviscross.com>